Refactor checklist functionality and style
