+++ components/InfoBox/HotspotDetails/Checklist/ChecklistCheck.js
@@ -1,10 +1,11 @@
-const ChecklistCheck = () => {
+const ChecklistCheck = ({ className }) => {
   return (
     <svg
       width="20"
       height="20"
       viewBox="0 0 20 20"
       fill="none"
+      className={className}
       xmlns="http://www.w3.org/2000/svg"
     >
       <path
